+++ dxrk/__init__.py
@@ -3,5 +3,5 @@
 
 from __future__ import annotations
 
-__version__ = "0.2.3"
+__version__ = "0.2.4"
 __all__ = ["__version__"]